add notification types file
